+++ inertia/pages/boats/show.vue
@@ -5,7 +5,6 @@ import BaseBreadcrumb from '~/components/base/BaseBreadcrumb.vue'
 import BaseButton from '~/components/base/BaseButton.vue'
 import BaseHeading from '~/components/base/BaseHeading.vue'
 import BaseTabs from '~/components/base/BaseTabs.vue'
-import BoatModeSwitcher from '~/components/boats/show/BoatModeSwitcher.vue'
 import BoatShowTabContent from '~/components/boats/show/BoatShowTabContent.vue'
 import { useT } from '~/composables/use_t'
 import type {
@@ -187,11 +186,7 @@ function openTasksTab() {
         </div>
       </div>
 
-      <!-- Mode switcher + Tab bar -->
-      <div class="flex flex-col gap-4 sm:flex-row sm:items-center sm:gap-6">
-        <BoatModeSwitcher :boat-id="boat.id" mode="management" />
-        <BaseTabs v-model="tab" :tabs="tabs" />
-      </div>
+      <BaseTabs v-model="tab" :tabs="tabs" />
     </header>
 
     <BoatShowTabContent